+++ src/onegov.core/onegov/core/tests/test_layout.py
@@ -1,6 +1,7 @@
-from datetime import datetime
-from sedate import replace_timezone
+from datetime import datetime, timedelta
+from sedate import replace_timezone, utcnow
 from onegov.core.layout import Layout
+from onegov.core.utils import Bunch
 
 
 def test_chunks():
@@ -33,3 +34,9 @@ def test_format_number():
     assert layout.format_number(1000) == "1'000"
     assert layout.format_number(1000.00) == "1'000.00"
     assert layout.format_number(1000.00, 0) == "1'000"
+
+
+def test_relative_date():
+    layout = Layout(model=object(), request=Bunch(locale='en'))
+    text = layout.format_date(utcnow() - timedelta(seconds=60 * 5), 'relative')
+    assert text == '5 minutes ago'